4th year/Cloud/VK Cloud/notes_app_v2/requirements.txt
Flask
flask_login
SQLAlchemy
psycopg2-binary
python-dotenv